bioinformatics/pyproject.toml
[tool.poetry]
name = "bioinformatics"
version = "0.1.0"
description = ""
authors = ["Nicolas J. Dowdy <njdowdy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
biopython = "^1.79"
pydantic = "^1.9.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^22.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
